+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [6.7.0] - 2022-12-03
+
+### Changed
+
+* The `DatabaseAdapter` now supports League's [checksum](https://flysystem.thephpleague.com/docs/usage/checksums/) operation. [#121](https://github.com/aedart/athenaeum/issues/121).
+
+### Removed
+
+* `throw` option for flysystem database connection. This option was never used.
+
+### Deprecated
+
+* `language()` method in `SearchProcessor`. This method has no effect on the search processor, nor its underlying `SearchFilter`. [#125](https://github.com/aedart/athenaeum/issues/125).
+
 ## [6.6.0] - 2022-11-28
 
 ### Added
@@ -877,7 +891,8 @@ It will highjack the `app` binding, which will cause your application to behave 
 
 * Please review commits on [GitHub](https://github.com/aedart/athenaeum/commits/master)
 
-[Unreleased]: https://github.com/aedart/athenaeum/compare/6.6.0...HEAD
+[Unreleased]: https://github.com/aedart/athenaeum/compare/6.7.0...HEAD
+[6.7.0]: https://github.com/aedart/athenaeum/compare/6.6.0...6.7.0
 [6.6.0]: https://github.com/aedart/athenaeum/compare/6.5.2...6.6.0
 [6.5.2]: https://github.com/aedart/athenaeum/compare/6.5.1...6.5.2
 [6.5.1]: https://github.com/aedart/athenaeum/compare/6.5.0...6.5.1